requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
py-cord==2.5.0
python-dotenv==1.0.1
six==1.16.0
speedruncompy==0.4.0
yarl==1.9.4
